Make `request` API accept a `BroadcastConfig`
By default, all broadcasts are pushed out into
each and every connected endpoint no matter independent
of the question if these processes are listening
for these events or not.

A `BroadcastConfig` allows reducing the scope
of a broadcast which can be used as a way to tame
resource consumption. E.g. as a developer we can
decide to broadcast specific events only to
specific endpoints if we know that these should
only be consumed by that endpoint anyway.

So far the `request` API did not accept a
`BroadcastConfig` which means these API calls
could not be optimized in that way.

This change makes it so that the `request` API
accepts a `BroadcastConfig` as well.