Fix error when future is cancelled after being set
It is rare, but possible, for a response to come in and cause the
future to be set and removed from Endpoint._futures just before the
future is canceled, which triggers _remove_cancelled_future and the
del Endpoint._futures[] found there.

This PR turns the del into a pop(, None) If the future was already
removed for some reason then _remove_cancelled_future shouldn't care.