Get rid of central coordinator topology
Previous to this, Lahja used a central
coordinator process where each event had
to travel through.

This gets rid of the central coordinator
and instead connects each Endpoint directly
with each other Endpoint that it wants
to push messages into.

It also gets rid of additional threading
overhead which all in all leads to a major
performance improvement.